Filter existing matches for interface speed
This commit updates the interface to filter down existing matches
instead of the entire choice set when new characters are inputted.

I'm not yet familiar enough with the lifetime system to figure out how
to accomplish this with the existing Choice::text() method, so I made
the struct field public instead.

Unfortunately, switching from a tuple struct to a regular old struct
resulted in a slowdown in benchmarks.