Use Iterator#size_hint() when allocating bonus vector
This commit updates the implementation of compute_bonus to make use of
the `size_hint` method in Iterator to determine the bonus vec's capacity.

This results in a 11% speed boost, according to `cargo bench`.

My understanding from some additional testing is that `size_hint` is not
always correct for non-Latin characters. As the derived value is only
used to allocate space for a vector, I'm comfortable with this tradeoff
for now, as it just means that some strings _may_ result in additional
allocations.