Remove unnecessary AsciiExt imports
These are not required on newer stable releases of the Rust compiler,
and generate a compile-time warning, even with our existing
`#[allow(unused_exports)]` pragma.

However, as this is now supported in stable releases, and this project
is still pre-1.0, now's as good a time as any to drop these imports
entirely.