Interface::run returns a Cow<'a>
Changed `Interface::run` to return a `Cow` with the lifetime of the borrowed slice of strings. 

The `Owned` case occurs when the search does not match anything. In that case, the search string is taken out of the `Interface` and replaced by an empty string.

Using a `Cow` with the lifetime of the given strings allows to drop the `Interface` struct after a run, without needing to clone the returned `&str` to keep the result. It also allows the caller to know whether the result is from the given list, or comes from a search that had no match.